refactor: move rate types, rename to utils and accept params
