+++ src/console/src/impls.rs
@@ -4,7 +4,7 @@ use crate::types::core::CommitProposalError;
 use crate::types::ledger::Payment;
 use crate::types::state::{
     Fee, Fees, HeapState, MissionControl, Proposal, ProposalKey, ProposalStatus, ProposalType,
-    Rates, State,
+    Rate, Rates, State,
 };
 use candid::Principal;
 use ic_cdk::api::time;
@@ -12,7 +12,7 @@ use ic_stable_structures::storable::Bound;
 use ic_stable_structures::Storable;
 use junobuild_shared::constants::INITIAL_VERSION;
 use junobuild_shared::rate::constants::DEFAULT_RATE_CONFIG;
-use junobuild_shared::rate::types::{Rate, RateTokens};
+use junobuild_shared::rate::types::RateTokens;
 use junobuild_shared::serializers::{deserialize_from_bytes, serialize_to_bytes};
 use junobuild_shared::types::core::Hash;
 use junobuild_shared::types::state::Version;
