+++ src/libs/shared/src/rate/mod.rs
@@ -1,3 +1,3 @@
 pub mod constants;
-pub mod quota;
 pub mod types;
+pub mod utils;
