+++ src/libs/shared/src/rate/types.rs
@@ -3,12 +3,6 @@ use candid::{CandidType, Deserialize};
 use serde::Serialize;
 
 #[derive(Default, CandidType, Serialize, Deserialize, Clone)]
-pub struct Rate {
-    pub tokens: RateTokens,
-    pub config: RateConfig,
-}
-
-#[derive(Default, CandidType, Serialize, Deserialize, Clone)]
 pub struct RateTokens {
     pub tokens: u64,
     pub updated_at: Timestamp,
